third_party/blink/web_tests/external/wpt/html/semantics/popups/popup-light-dismiss.tentative.html

<!DOCTYPE html>
<meta charset="utf-8" />
<title>Popup light dismiss behavior</title>
<link rel="author" href="mailto:masonf@chromium.org">
<link rel=help href="https://open-ui.org/components/popup.research.explainer">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<button id=b1 togglepopup='p1'>Popup 1</button>
<button id=p1anchor>Popup1 anchor (no action)</button>
<span id=outside>Outside all popups</span>
<div popup=popup id=p1 anchor=p1anchor>
  <span id=inside1>Inside popup 1</span>
  <button id=b2 togglepopup='p2'>Popup 2</button>
  <span id=inside1after>Inside popup 1 after button</span>
</div>
<div popup=popup id=p2 anchor=b2>
  <span id=inside2>Inside popup 2</span>
</div>

<button id=b3 togglepopup=p3>Popup 3 - button 3
  <div popup=popup id=p4>Inside popup 4</div>
</button>
<div popup=popup id=p3>Inside popup 3</div>
<button id=b4 popup=p3>Popup 3 - button 4
  <div popup=popup id=p5>Inside popup 5</div>
</button>

<div popup=popup id=p6>Inside popup 6
  <div style="height:2000px;background:lightgreen"></div>
  Bottom of popup6
</div>
<button popup=p6>Popup 6</button>

<my-element id="myElement">
  <template shadowroot="open">
    <button id=b7 onclick='showPopup7()'>Popup7</button>
    <div popup=popup id=p7 anchor=b7 style="top: 100px;">
      <p>Popup content.</p>
      <input id="inside7" type="text" placeholder="some text">
    </div>
  </template>
</my-element>

<div popup=popup id=p8 anchor=p8anchor>
  <button>Button</button>
  <span id=inside8after>Inside popup 8 after button</span>
</div>
<button id=p8anchor>Popup8 anchor (no action)</button>

<div popup=popup id=p9>
  <button>Button</button>
  <span id=inside9after>Inside popup 9 after button</span>
</div>
<button id=b9after togglepopup='p9'>Popup 9</button>


<style>
  #p1 { top:50px; }
  #p2 { top:50px; left:250px; }
  #p3 { top:150px;}
  #p4 { top:150px; left:250px;}
  #p5 { top:250px;}
  #p6 {
    top: 50px;
    left:150px;
    width: 300px;
    height: 300px;
    overflow-y: scroll;
  }
  popup { border: 5px solid red; }
</style>


<script>
  function clickOn(element) {
    const actions = new test_driver.Actions();
    return actions.pointerMove(0, 0, {origin: element})
      .pointerDown({button: actions.ButtonType.LEFT})
      .pointerUp({button: actions.ButtonType.LEFT})
      .send();
  }

  const popup1 = document.querySelector('#p1');
  const button1 = document.querySelector('#b1');
  const popup1anchor = document.querySelector('#p1anchor');
  const inside1After = document.querySelector('#inside1after');
  const popup2 = document.querySelector('#p2');
  const outside = document.querySelector('#outside');
  const inside1 = document.querySelector('#inside1');
  const inside2 = document.querySelector('#inside2');
  const button3 = document.querySelector('#b3');
  const button4 = document.querySelector('#b4');
  const popup3 = document.querySelector('#p3');
  const popup4 = document.querySelector('#p4');
  const popup5 = document.querySelector('#p5');
  const popup6 = document.querySelector('#p6');
  const button7 = document.querySelector('#myElement').shadowRoot.querySelector('#b7');
  const popup7 = document.querySelector('#myElement').shadowRoot.querySelector('#p7');
  const inside7 = document.querySelector('#myElement').shadowRoot.querySelector('#inside7');

  function showPopup7() {
    popup7.showPopup();
  }
  let popup1HideCount = 0;
  popup1.addEventListener('hide',(e) => {
    ++popup1HideCount;
    e.preventDefault(); // 'hide' should not be cancellable.
  });
  let popup2HideCount = 0;
  popup2.addEventListener('hide',(e) => {
    ++popup2HideCount;
    e.preventDefault(); // 'hide' should not be cancellable.
  });
  function spinEventLoop() {
    return new Promise(resolve => requestAnimationFrame(resolve));
  }

  promise_test(async () => {
    assert_false(popup1.matches(':popup-open'));
    popup1.showPopup();
    let p1HideCount = popup1HideCount;
    assert_true(popup1.matches(':popup-open'));
    await clickOn(outside);
    assert_false(popup1.matches(':popup-open'));
    assert_equals(popup1HideCount,p1HideCount+1);
  },'Clicking outside a popup will dismiss the popup');

  promise_test(async () => {
    assert_false(popup1.matches(':popup-open'));
    popup1.showPopup();
    p1HideCount = popup1HideCount;
    await clickOn(inside1);
    assert_true(popup1.matches(':popup-open'));
    assert_equals(popup1HideCount,p1HideCount);
    popup1.hidePopup();
  },'Clicking inside a popup does not close that popup');

  promise_test(async () => {
    assert_false(popup1.matches(':popup-open'));
    popup1.showPopup();
    await clickOn(inside1After);
    assert_true(popup1.matches(':popup-open'));
    await new test_driver.send_keys(document.body,'\uE004'); // Tab
    assert_equals(document.activeElement,button3,'Focus should move to a button outside the popup');
    assert_false(popup1.matches(':popup-open'));
  },'Moving focus outside the popup will dismiss the popup');

  promise_test(async () => {
    popup1.showPopup();
    popup2.showPopup();
    await spinEventLoop();
    p1HideCount = popup1HideCount;
    let p2HideCount = popup2HideCount;
    await clickOn(inside2);
    assert_true(popup1.matches(':popup-open'),'popup1 should be open');
    assert_true(popup2.matches(':popup-open'),'popup2 should be open');
    assert_equals(popup1HideCount,p1HideCount,'popup1');
    assert_equals(popup2HideCount,p2HideCount,'popup2');
    popup1.hidePopup();
    popup2.hidePopup();
  },'Clicking inside a child popup shouldn\'t close either popup');

  promise_test(async () => {
    popup1.showPopup();
    popup2.showPopup();
    await spinEventLoop();
    p1HideCount = popup1HideCount;
    p2HideCount = popup2HideCount;
    await clickOn(inside1);
    assert_true(popup1.matches(':popup-open'));
    assert_equals(popup1HideCount,p1HideCount);
    assert_false(popup2.matches(':popup-open'));
    assert_equals(popup2HideCount,p2HideCount+1);
    popup1.hidePopup();
  },'Clicking inside a parent popup should close child popup');

  promise_test(async () => {
    popup1.showPopup();
    assert_true(popup1.matches(':popup-open'));
    await spinEventLoop();
    p1HideCount = popup1HideCount;
    await clickOn(popup1anchor);
    assert_true(popup1.matches(':popup-open'),'popup1 not open');
    assert_equals(popup1HideCount,p1HideCount);
    popup1.hidePopup(); // Cleanup
    assert_false(popup1.matches(':popup-open'));
  },'Clicking on anchor element (that isn\'t an invoking element) shouldn\'t close its popup');

  promise_test(async () => {
    await clickOn(button3);
    assert_true(popup3.matches(':popup-open'),'invoking element should open popup');
    popup4.showPopup();
    assert_true(popup4.matches(':popup-open'));
    assert_true(popup3.matches(':popup-open'));
    popup3.hidePopup(); // Cleanup
    assert_false(popup3.matches(':popup-open'));
  },'An invoking element should be part of the ancestor chain');

  promise_test(async () => {
    await clickOn(button3);
    assert_true(popup3.matches(':popup-open'));
    assert_false(popup4.matches(':popup-open'));
    assert_false(popup5.matches(':popup-open'));
    popup5.showPopup();
    assert_true(popup5.matches(':popup-open'));
    assert_false(popup3.matches(':popup-open'));
    popup5.hidePopup();
  },'An invoking element that was not used to invoke the popup should NOT be part of the ancestor chain');

  promise_test(async () => {
    popup1.showPopup();
    popup2.showPopup(); // Popup1 is an ancestral element for popup2.
    assert_true(popup1.matches(':popup-open'));
    assert_true(popup2.matches(':popup-open'));
    const drag_actions = new test_driver.Actions();
    // Drag *from* popup2 *to* popup1 (its ancestor).
    await drag_actions.pointerMove(0,0,{origin: popup2})
      .pointerDown({button: drag_actions.ButtonType.LEFT})
      .pointerMove(0,0,{origin: popup1})
      .pointerUp({button: drag_actions.ButtonType.LEFT})
      .send();
    assert_true(popup1.matches(':popup-open'),'popup1 should be open');
    assert_true(popup2.matches(':popup-open'),'popup1 should be open');
    popup1.hidePopup();
    assert_false(popup2.matches(':popup-open'));
  },'Dragging from an open popup outside an open popup should leave the popup open');

  promise_test(async () => {
    popup6.showPopup();
    assert_equals(popup6.scrollTop,0,'popup6 should start non-scrolled');
    await new test_driver.Actions()
       .scroll(0, 0, 0, 50, {origin: popup6})
       .send();
    assert_true(popup6.matches(':popup-open'),'popup6 should stay open');
    assert_equals(popup6.scrollTop,50,'popup6 should be scrolled');
    popup6.hidePopup();
  },'Scrolling within a popup should not close the popup');

  promise_test(async () => {
    button7.click();
    assert_true(popup7.matches(':popup-open'),'invoking element should open popup');
    inside7.click();
    assert_true(popup7.matches(':popup-open'));
    popup7.hidePopup();
  },'Clicking inside a shadow DOM popup does not close that popup');

  promise_test(async () => {
    button7.click();
    inside7.click();
    assert_true(popup7.matches(':popup-open'));
    await clickOn(outside);
    assert_false(popup7.matches(':popup-open'));
  },'Clicking outside a shadow DOM popup should close that popup');

  promise_test(async () => {
    const popup8 = document.querySelector('#p8');
    const inside8After = document.querySelector('#inside8after');
    const popup8Anchor = document.querySelector('#p8anchor');
    assert_false(popup8.matches(':popup-open'));
    popup8.showPopup();
    await clickOn(inside8After);
    assert_true(popup8.matches(':popup-open'));
    await new test_driver.send_keys(document.body,'\uE004'); // Tab
    assert_equals(document.activeElement,popup8Anchor,'Focus should move to the anchor element');
    assert_true(popup8.matches(':popup-open'),'popup should stay open');
    popup8.hidePopup();
  },'Moving focus back to the anchor element should not dismiss the popup');

  promise_test(async () => {
    const popup9 = document.querySelector('#p9');
    const inside9After = document.querySelector('#inside9after');
    const popup9Invoker = document.querySelector('#b9after');
    assert_false(popup9.matches(':popup-open'));
    popup9Invoker.click(); // Trigger via the button
    await clickOn(inside9After);
    assert_true(popup9.matches(':popup-open'));
    await new test_driver.send_keys(document.body,'\uE004'); // Tab
    assert_equals(document.activeElement,popup9Invoker,'Focus should move to the invoking element');
    assert_true(popup9.matches(':popup-open'),'popup should stay open');
    popup9.hidePopup();
  },'Moving focus back to the active trigger element should not dismiss the popup');

  promise_test(async () => {
    const popup9 = document.querySelector('#p9');
    const inside9After = document.querySelector('#inside9after');
    const popup9Invoker = document.querySelector('#b9after');
    assert_false(popup9.matches(':popup-open'));
    popup9.showPopup(); // Trigger directly
    await clickOn(inside9After);
    assert_true(popup9.matches(':popup-open'));
    await new test_driver.send_keys(document.body,'\uE004'); // Tab
    assert_equals(document.activeElement,popup9Invoker,'Focus should move to the invoking element');
    assert_false(popup9.matches(':popup-open'),'popup should close - triggering element was not used');
  },'Moving focus back to the inactive trigger element should dismiss the popup');

</script>
